hal: renesas: Restructure source codes
Restructure source code for easy maintaining and adding new MCU series
