+++ drivers/ra/fsp/src/bsp/mcu/all/bsp_register_protection.h
