+++ zephyr/ra/CMakeLists.txt
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+
+add_subdirectory(ra_cfg)
+zephyr_include_directories(portable)
