+++ zephyr/ra/ra_cfg/CMakeLists.txt
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_include_directories(fsp_cfg)
+zephyr_include_directories(fsp_cfg/bsp/${CONFIG_SOC_SERIES})
